Bring command bar for container scenario in parity with code command bar.
